+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Overriding built-in components in Markdown.
 
+- Removed duplicate color `baseHeading`.
+
 ## [0.31.0] - 2020-12-17
 
 - Added `name` prop to Switcher.